New refactor.
