+++ butterfree/load/writers/online_feature_store_writer.py
@@ -1,8 +1,10 @@
 """Holds the Online Feature Store writer class."""
-from typing import Any, List
+import os
+from typing import Any, List, Optional
 
 from pyspark.sql import DataFrame, Window
 from pyspark.sql.functions import col, row_number
+from pyspark.sql.streaming import StreamingQuery
 
 from butterfree.clients import CassandraClient, SparkClient
 from butterfree.configs.db import CassandraConfig
@@ -76,11 +78,12 @@ class OnlineFeatureStoreWriter(Writer):
         db_config=None,
         debug_mode: bool = False,
         write_to_entity: bool = False,
-        check_schema: Hook = None,
+        check_schema_hook: Hook = None,
     ):
-        super().__init__(debug_mode, write_to_entity)
+        super().__init__(debug_mode)
         self.db_config = db_config or CassandraConfig()
-        self.check_schema = check_schema
+        self.check_schema_hook = check_schema_hook
+        self.write_to_entity = write_to_entity
 
     @staticmethod
     def filter_latest(dataframe: DataFrame, id_columns: List[Any]) -> DataFrame:
@@ -112,21 +115,20 @@ class OnlineFeatureStoreWriter(Writer):
 
     def load(
         self, feature_set: FeatureSet, dataframe: DataFrame, spark_client: SparkClient,
-    ):
-        """Prepare the dataframe before it is saved to the Online Feature Store.
+    ) -> Optional[StreamingQuery]:
+        """Loads the latest data from a feature set into the Feature Store.
+
+        Feature Store could be Online or Historical.
 
         Args:
             feature_set: object processed with feature set metadata.
             dataframe: Spark dataframe containing data from a feature set.
             spark_client: client for Spark connections with external services.
 
+        If dataframe is streaming this temporary table will be updated in real time.
+
         Returns:
-            load_df: Dataframe ready to be saved.
-            db_config: Spark configuration for connect databases.
-            options(optional = None): All other string options.
-            database(optional = None): Database name where the dataframe will be saved.
-            table_name: Table name where the dataframe will be saved.
-            partition_by(optional = None): Partition column to use when writing.
+            Streaming handler if writing streaming df, None otherwise.
 
         """
         table_name = feature_set.entity if self.write_to_entity else feature_set.name
@@ -138,24 +140,31 @@ class OnlineFeatureStoreWriter(Writer):
             password=self.db_config.password,
         )
 
-        if not self.check_schema:
-            self.check_schema = CassandraTableSchemaCompatibilityHook(
-                cassandra_client, table_name
+        self.check_schema(cassandra_client, table_name)
+        self.run_pre_hooks(dataframe)
+
+        if dataframe.isStreaming:
+            if self.debug_mode:
+                return self.write_in_debug_mode(
+                    table_name=f"online_feature_store__{table_name}",
+                    dataframe=dataframe,
+                    spark_client=spark_client,
+                )
+            return self._write_stream(
+                feature_set=feature_set,
+                dataframe=dataframe,
+                spark_client=spark_client,
+                table_name=table_name,
+                db_config=self.db_config,
             )
 
-        self.add_pre_hook(self.check_schema)
-
-        df = (
-            dataframe
-            if dataframe.isStreaming
-            else self.filter_latest(
-                dataframe=dataframe, id_columns=feature_set.keys_columns
-            )
+        latest_df = self.filter_latest(
+            dataframe=dataframe, id_columns=feature_set.keys_columns
         )
 
         options = self.db_config.get_options(table=table_name)
 
-        return (df, self.db_config, options, None, table_name, None)
+        self.write(spark_client, latest_df, table_name, self.db_config, options, None)
 
     def validate(self, feature_set: FeatureSet, dataframe, spark_client: SparkClient):
         """Calculate dataframe rows to validate data into Feature Store.
@@ -176,6 +185,21 @@ class OnlineFeatureStoreWriter(Writer):
         # TODO how to run data validations when a database has concurrent writes.
         pass
 
+    def check_schema(self, client, table_name, database=None):
+        """Instantiate the schema check hook and add it to the list of pre hooks.
+
+        Args:
+            client: client for Spark or Cassandra connections with external services.
+            table_name: table name where the dataframe will be saved.
+            database: database name where the dataframe will be saved.
+        """
+        if not self.check_schema_hook:
+            self.check_schema_hook = CassandraTableSchemaCompatibilityHook(
+                client, table_name
+            )
+
+        self.add_pre_hook(self.check_schema_hook)
+
     def get_db_schema(self, feature_set: FeatureSet):
         """Get desired database schema.
 
@@ -188,3 +212,33 @@ class OnlineFeatureStoreWriter(Writer):
         """
         db_schema = self.db_config.translate(feature_set.get_schema())
         return db_schema
+
+    def _write_stream(
+        self,
+        feature_set: FeatureSet,
+        dataframe: DataFrame,
+        spark_client: SparkClient,
+        table_name: str,
+        db_config,
+    ) -> StreamingQuery:
+        """Writes the dataframe in streaming mode."""
+        checkpoint_folder = (
+            f"{feature_set.name}__on_entity" if self.write_to_entity else table_name
+        )
+        checkpoint_path = (
+            os.path.join(
+                db_config.stream_checkpoint_path, feature_set.entity, checkpoint_folder,
+            )
+            if db_config.stream_checkpoint_path
+            else None
+        )
+        streaming_handler = spark_client.write_stream(
+            dataframe,
+            processing_time=db_config.stream_processing_time,
+            output_mode=db_config.stream_output_mode,
+            checkpoint_path=checkpoint_path,
+            format_=db_config.format_,
+            mode=db_config.mode,
+            **db_config.get_options(table=table_name),
+        )
+        return streaming_handler
