+++ tests/unit/butterfree/load/writers/test_online_feature_store_writer.py
@@ -1,5 +1,11 @@
-import pytest
+from unittest.mock import Mock
 
+import pytest
+from pyspark.sql import DataFrame
+from pyspark.sql.streaming import StreamingQuery
+
+from butterfree.clients import SparkClient
+from butterfree.configs.db import CassandraConfig
 from butterfree.load.writers import OnlineFeatureStoreWriter
 
 
@@ -57,21 +63,29 @@ class TestOnlineFeatureStoreWriter:
     ):
         # with
         spark_client = mocker.stub("spark_client")
+        spark_client.write_dataframe = mocker.stub("write_dataframe")
         writer = OnlineFeatureStoreWriter(cassandra_config)
 
+        writer.run_pre_hooks = Mock()
+        writer.run_pre_hooks.return_value = feature_set_dataframe
+
         # when
-        result_df, db_config, options, database, table_name, partition_by = writer.load(
-            feature_set, feature_set_dataframe, spark_client
-        )
+        writer.load(feature_set, feature_set_dataframe, spark_client)
 
         assert sorted(latest_feature_set_dataframe.collect()) == sorted(
-            result_df.collect()
+            spark_client.write_dataframe.call_args[1]["dataframe"].collect()
         )
-        assert writer.db_config == db_config
+        assert (
+            writer.db_config.mode == spark_client.write_dataframe.call_args[1]["mode"]
+        )
+        assert (
+            writer.db_config.format_
+            == spark_client.write_dataframe.call_args[1]["format_"]
+        )
         # assert if all additional options got from db_config
         # are in the called args in write_dataframe
         assert all(
-            item in options.items()
+            item in spark_client.write_dataframe.call_args[1].items()
             for item in writer.db_config.get_options(table=feature_set.name).items()
         )
 
@@ -80,3 +94,108 @@ class TestOnlineFeatureStoreWriter:
         schema = writer.get_db_schema(test_feature_set)
 
         assert schema == expected_schema
+
+    def test_write_in_debug_and_stream_mode(
+        self, feature_set, spark_session, mocked_stream_df
+    ):
+        # arrange
+        spark_client = SparkClient()
+
+        writer = OnlineFeatureStoreWriter(debug_mode=True)
+        writer.run_pre_hooks = Mock()
+        writer.run_pre_hooks.return_value = mocked_stream_df
+
+        # act
+        handler = writer.load(
+            feature_set=feature_set,
+            dataframe=mocked_stream_df,
+            spark_client=spark_client,
+        )
+
+        # assert
+        mocked_stream_df.format.assert_called_with("memory")
+        mocked_stream_df.queryName.assert_called_with(
+            f"online_feature_store__{feature_set.name}"
+        )
+        assert isinstance(handler, StreamingQuery)
+
+    @pytest.mark.parametrize("has_checkpoint", [True, False])
+    def test_write_stream(self, feature_set, has_checkpoint, monkeypatch):
+        # arrange
+        spark_client = SparkClient()
+        spark_client.write_stream = Mock()
+        spark_client.write_dataframe = Mock()
+        spark_client.write_stream.return_value = Mock(spec=StreamingQuery)
+
+        dataframe = Mock(spec=DataFrame)
+        dataframe.isStreaming = True
+
+        if has_checkpoint:
+            monkeypatch.setenv("STREAM_CHECKPOINT_PATH", "test")
+
+        cassandra_config = CassandraConfig(keyspace="feature_set")
+        target_checkpoint_path = (
+            "test/entity/feature_set_test"
+            if cassandra_config.stream_checkpoint_path
+            else None
+        )
+
+        writer = OnlineFeatureStoreWriter(cassandra_config)
+        writer.run_pre_hooks = Mock()
+        writer.run_pre_hooks.return_value = dataframe
+
+        # act
+        stream_handler = writer.load(feature_set, dataframe, spark_client)
+
+        # assert
+        assert isinstance(stream_handler, StreamingQuery)
+        spark_client.write_stream.assert_any_call(
+            dataframe,
+            processing_time=cassandra_config.stream_processing_time,
+            output_mode=cassandra_config.stream_output_mode,
+            checkpoint_path=target_checkpoint_path,
+            format_=cassandra_config.format_,
+            mode=cassandra_config.mode,
+            **cassandra_config.get_options(table=feature_set.name),
+        )
+        spark_client.write_dataframe.assert_not_called()
+
+    def test_write_stream_on_entity(self, feature_set, monkeypatch):
+        """Test write method with stream dataframe and write_to_entity enabled.
+
+        The main purpose of this test is assert the correct setup of stream checkpoint
+        path and if the target table name is the entity.
+
+        """
+
+        # arrange
+        spark_client = SparkClient()
+        spark_client.write_stream = Mock()
+        spark_client.write_stream.return_value = Mock(spec=StreamingQuery)
+
+        dataframe = Mock(spec=DataFrame)
+        dataframe.isStreaming = True
+
+        feature_set.entity = "my_entity"
+        feature_set.name = "my_feature_set"
+        monkeypatch.setenv("STREAM_CHECKPOINT_PATH", "test")
+        target_checkpoint_path = "test/my_entity/my_feature_set__on_entity"
+
+        writer = OnlineFeatureStoreWriter(write_to_entity=True)
+        writer.run_pre_hooks = Mock()
+        writer.run_pre_hooks.return_value = dataframe
+
+        # act
+        stream_handler = writer.load(feature_set, dataframe, spark_client)
+
+        # assert
+        assert isinstance(stream_handler, StreamingQuery)
+        spark_client.write_stream.assert_called_with(
+            dataframe,
+            processing_time=writer.db_config.stream_processing_time,
+            output_mode=writer.db_config.stream_output_mode,
+            checkpoint_path=target_checkpoint_path,
+            format_=writer.db_config.format_,
+            mode=writer.db_config.mode,
+            **writer.db_config.get_options(table="my_entity"),
+        )
